Create a package to contain the plugin
