+++ sublime_isort/__init__.py
@@ -1,0 +1,3 @@
+from .commands import IsortCommand
+
+__all__ = ['IsortCommand']
